0206/study02/app3/pyproject.toml
[project]
name = "app3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-mail[standard]>=1.6.1",
    "fastapi[standard]>=0.128.2",
    "pydantic-settings[standard]>=2.12.0",
]
